+++ appletree/randgen.py
@@ -7,7 +7,6 @@ import jax
 from jax import numpy as jnp
 import numpy as np
 from jax import random, lax, jit, vmap
-from numpyro.distributions.util import _binomial_dispatch as _binomial_dispatch_numpyro
 from scipy.interpolate import interp1d
 
 from appletree.utils import exporter
@@ -152,7 +151,7 @@ def truncate_normal(key, mean, std, vmin=None, vmax=None, shape=()):
 
     """
     key, rvs = normal(key, mean, std, shape=shape)
-    rvs = jnp.clip(rvs, a_min=vmin, a_max=vmax)
+    rvs = jnp.clip(rvs, vmin, vmax)
     return key, rvs.astype(FLOAT)
 
 
@@ -246,6 +245,8 @@ if hasattr(random, "binomial"):
         return key, rvs.astype(INT)
 
 else:
+    from numpyro.distributions.util import _binomial_dispatch as _binomial_dispatch_numpyro
+
     warn("random.binomial is not available, using numpyro's _binomial_dispatch instead.")
     if os.environ.get("DO_NOT_USE_APPROX_IN_BINOM") is None:
         ALWAYS_USE_NORMAL_APPROX_IN_BINOM = True
@@ -277,7 +278,7 @@ else:
             q = 1.0 - p
             mean = n * p
             std = jnp.sqrt(n * p * q)
-            rvs = jnp.clip(random.normal(seed) * std + mean, a_min=0.0, a_max=n)
+            rvs = jnp.clip(random.normal(seed) * std + mean, 0.0, n)
             return rvs.round().astype(INT)
 
         def _binomial_dispatch(seed, p, n):
@@ -312,21 +313,22 @@ else:
 @export
 @partial(jit, static_argnums=(3,))
 def negative_binomial(key, p, n, shape=()):
-    """Negative binomial distribution random sampler. Using Gamma–Poisson mixture.
+    """Negative binomial distribution random sampler.
 
-    Args:
-        key: seed for random generator.
-        p: <jnp.array>-like probability of a single success in negative binomial distribution.
-        n: <jnp.array>-like number of successes in negative binomial distribution.
-        shape: output shape.
-            If not given, output has shape jnp.broadcast_shapes(jnp.shape(p), jnp.shape(n)).
+    Using Gamma–Poisson mixture.
+        Args:
+            key: seed for random generator.
+            p: <jnp.array>-like probability of a single success in negative binomial distribution.
+            n: <jnp.array>-like number of successes in negative binomial distribution.
+            shape: output shape.
+                If not given, output has shape jnp.broadcast_shapes(jnp.shape(p), jnp.shape(n)).
 
-    Returns:
-        an updated seed, random variables.
+        Returns:
+            an updated seed, random variables.
 
-    References:
-        1. https://en.wikipedia.org/wiki/Negative_binomial_distribution#Gamma%E2%80%93Poisson_mixture  # noqa
-        2. https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.nbinom.html
+        References:
+            1. https://en.wikipedia.org/wiki/Negative_binomial_distribution#Gamma%E2%80%93Poisson_mixture  # noqa
+            2. https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.nbinom.html
 
     """